Refact: Rename 'Video' components to 'Watch'
